chore: (#730) response 타입은 api 파일로 이동
